Add files via upload
In an augmented version of the Needleman-Wunsch sequence alignment algorithm, these Perl programs collect user inputs for the cost of match, mismatch, and indels, as well as the two DNA sequences of interest. The first modification calculates the maximum alignment score. Gap_modification is made to work in the case of gaps, and LCS_modification has fixed cost values to ensure it works as the least common subsequence algorithm.